server/pyproject.toml
[project]
name = "election-mud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "feedparser>=6.0.12",
    "google-genai>=1.68.0",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "openai>=2.29.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
]
